requirements.txt
streamlit == 1.19
numpy
pandas
AgGrid
deep_translator